requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.0
asyncio==3.4.3
asyncpg==0.25.0
attrs==21.2.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
discord.py==1.7.3
ffmpeg==1.4
idna==2.8
idna-ssl==1.1.0
imutils==0.5.1
kiwisolver==1.0.1
lxml==4.4.1
matplotlib==3.0.0
multidict==5.2.0
numpy==1.15.2
opencv-python==3.4.3.18
pafy==0.5.5
pychalk==2.0.0
pycparser==2.19
PyNaCl==1.0.1
pyparsing==2.2.2
PySocks==1.7.1
python-dateutil==2.7.3
requests==2.22.0
requests-futures==1.0.0
rethinkdb==2.3.0.post6
scipy==1.1.0
six==1.11.0
soupsieve==1.9.5
stem==1.7.1
torrequest==0.1.0
typing-extensions==4.0.1
urllib3==1.25.6
utils==0.9.0
websockets==3.4
yarl==1.7.2
youtube-dl==2018.10.5